Recursivly discover Gemfile
Rakefiles can reside in subdirectories, with the Gemfile being located
at the project root, a good example for such a project is rails, where
each subdirectory has a Rakefile appropriate for its subproject but the
Gemfile is located all the way up at the root of rails itself.

A structure like this

```
~/sampledir $ tree
.
├── Gemfile
└── subdir
    └── Rakefile
```

In this case rake needs to still be run using bundle exec to make sure
the proper version of rake is used. To make this work the Rakefile
searched from the current rake root upwards.